+++ src/tests/XResponseTimeMW.Tests/GlobalUsings.cs
@@ -1,0 +1,6 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using System;
+global using System.Linq;
+global using System.Net;
+global using System.Net.Http;
+global using System.Threading.Tasks;